refactor Tracing file paths (#1664)
* refactor Tracing file paths

* fix references
