+++ BASE/Test/Microsoft.ApplicationInsights.Test/Microsoft.ApplicationInsights.Tests/Extensibility/Implementation/Tracing/DiagnosticsListenerTest.cs
@@ -4,6 +4,7 @@ namespace Microsoft.ApplicationInsights.Extensibility.Implementation.Tracing
     using System;
     using System.Collections.Generic;
     using System.Diagnostics.Tracing;
+    using Microsoft.ApplicationInsights.Extensibility.Implementation.Tracing.DiagnosticsModule;
     using Microsoft.VisualStudio.TestTools.UnitTesting;
     using Tracing.Mocks;
