+++ BASE/Test/Microsoft.ApplicationInsights.Test/Microsoft.ApplicationInsights.Tests/Extensibility/Implementation/Tracing/Mocks/DiagnoisticsEventThrottlingManagerMock.cs
@@ -6,6 +6,8 @@
 // <summary></summary>
 // -----------------------------------------------------------------------
 
+using Microsoft.ApplicationInsights.Extensibility.Implementation.Tracing.DiagnosticsModule;
+
 namespace Microsoft.ApplicationInsights.Extensibility.Implementation.Tracing.Mocks
 {
     internal class DiagnoisticsEventThrottlingManagerMock : IDiagnoisticsEventThrottlingManager
