+++ BASE/Test/Microsoft.ApplicationInsights.Test/Microsoft.ApplicationInsights.Tests/Extensibility/Implementation/Tracing/Mocks/DiagnoisticsEventThrottlingMock.cs
@@ -9,6 +9,7 @@
 namespace Microsoft.ApplicationInsights.Extensibility.Implementation.Tracing.Mocks
 {
     using System.Collections.Generic;
+    using Microsoft.ApplicationInsights.Extensibility.Implementation.Tracing.DiagnosticsModule;
 
     internal class DiagnoisticsEventThrottlingMock : IDiagnoisticsEventThrottling
     {
