+++ BASE/src/Microsoft.ApplicationInsights/Extensibility/Implementation/Tracing/BaseDefaultHeartbeatPropertyProvider.cs
