+++ BASE/src/Microsoft.ApplicationInsights/Extensibility/Implementation/Tracing/FileDiagnosticsTelemetryModule.cs
@@ -1,4 +1,4 @@
-﻿namespace Microsoft.ApplicationInsights.Extensibility.Implementation.Tracing.SelfDiagnostics
+﻿namespace Microsoft.ApplicationInsights.Extensibility.Implementation.Tracing
 {
 #if !NETSTANDARD1_3
 
@@ -9,6 +9,7 @@
     using System.IO;
     using System.Security;
     using System.Threading;
+    using Microsoft.ApplicationInsights.Extensibility.Implementation.Tracing.FileDiagnosticsModule;
 
     /// <summary>
     /// Diagnostics telemetry module for azure web sites.
