+++ BASE/src/Microsoft.ApplicationInsights/Extensibility/Implementation/Tracing/HeartbeatDefaultPayload.cs
