+++ BASE/src/Microsoft.ApplicationInsights/Extensibility/Implementation/Tracing/HeartbeatPropertyPayload.cs
