+++ BASE/src/Microsoft.ApplicationInsights/Extensibility/Implementation/Tracing/IHeartbeatDefaultPayloadProvider.cs
