+++ BASE/src/Microsoft.ApplicationInsights/Extensibility/Implementation/Tracing/IHeartbeatPropertyManager.cs
