+++ BASE/src/Microsoft.ApplicationInsights/Extensibility/Implementation/Tracing/IHeartbeatProvider.cs
